git-tfs-id: [http://uk-tfs02:8080/tfs/DefaultCollection]$/Erm.Core/Prerelease/BLCore;C17343

+++ Source/ApplicationServices/2Gis.Erm.BLCore.Aggregates/Common/Generics/ICreateAggregateRepository.cs
@@ -7,6 +7,6 @@ namespace DoubleGis.Erm.BLCore.Aggregates.Common.Generics
     public interface ICreateAggregateRepository<in TEntity> : IUnknownAggregateSpecificOperation<CreateIdentity>
         where TEntity : class, IEntity, IEntityKey
     {
-        int Create(TEntity entity);
+        long Create(TEntity entity);
     }
 }
